Small fix for optimize projection bin

+++ test/hzz2l2v/computeLimit/optimize_WideWidth.py
@@ -372,7 +372,7 @@ for signalSuffix in signalSuffixVec :
                      vals=line.split(' ')
                      for c in range(1, cutsH.GetYaxis().GetNbins()+3):
                         #FIXME FORCE INDEX TO BE 17 (Met>125GeV)
-                        Gcut[c-1].SetPoint(mi, 17, float(125));
+                        Gcut[c-1].SetPoint(mi, 17, float(130));
    #                     Gcut[c-1].SetPoint(mi, float(vals[0]), float(vals[c+1]));
                      mi+=1
                   for c in range(1, cutsH.GetYaxis().GetNbins()+3): Gcut[c-1].Set(mi);
@@ -382,7 +382,7 @@ for signalSuffix in signalSuffixVec :
                   for mtmp in SUBMASS:
                      for c in range(1, cutsH.GetYaxis().GetNbins()+3):
                         #FIXME FORCE INDEX TO BE 17 (Met>125GeV)
-                        Gcut[c-1].SetPoint(mi, 17, float(125));
+                        Gcut[c-1].SetPoint(mi, 17, float(130));
                      mi+=1
                   for c in range(1, cutsH.GetYaxis().GetNbins()+3): Gcut[c-1].Set(mi);
